fix: theme downloading

+++ symfony/src/Service/Theme/ThemeFilesService.php
@@ -11,6 +11,7 @@ use App\Service\Theme\Event\ConfigEditedEvent;
 use App\Service\Theme\Event\LangEditedEvent;
 use App\Service\Theme\Event\StylesEditedEvent;
 use App\Service\Theme\Event\TemplateEditedEvent;
+use App\Service\Theme\Exception\ThemeExportException;
 use App\Service\Theme\Exception\ThemeImportException;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Clock\ClockAwareTrait;
@@ -184,6 +185,10 @@ class ThemeFilesService
         };
     }
 
+    /**
+     * @throws ThemeExportException known errors
+     * @throws \RuntimeException for unexpected errors
+     */
     public function exportFilesToZip(Blog $blog): string
     {
         $tmpPath = tempnam(sys_get_temp_dir(), 'theme-export-');
@@ -195,7 +200,13 @@ class ThemeFilesService
             $zip = new \ZipArchive();
             $zip->open($tmpPath, \ZipArchive::OVERWRITE);
 
-            foreach ($this->getAllFilesOfBlog($blog) as $file) {
+            $files = $this->getAllFilesOfBlog($blog);
+
+            if (count($files) === 0) {
+                throw new ThemeExportException('No theme files found to export');
+            }
+
+            foreach ($files as $file) {
                 $folder = $file->getFolder();
                 $entryName = $folder === null ? $file->getName() : $folder->value . '/' . $file->getName();
                 $zip->addFromString($entryName, $file->getContent() ?? '');
@@ -210,7 +221,9 @@ class ThemeFilesService
 
             return $content;
         } finally {
-            unlink($tmpPath);
+            if (file_exists($tmpPath)) {
+                unlink($tmpPath);
+            }
         }
     }
